requirements.txt
adafruit-circuitpython-dht
fastapi[standard]
adafruit-circuitpython-gps
RPi.bme280